Implemented basic content indexing application.
